+++ back-end/src/main/java/br/edu/com/tg/manager/core/domain/entities/Professor.java
@@ -1,5 +1,6 @@
 package br.edu.com.tg.manager.core.domain.entities;
 
+import br.edu.com.tg.manager.core.domain.entities.enums.ProfessorRole;
 import br.edu.com.tg.manager.core.domain.exceptions.DomainException;
 
 /**
