+++ back-end/src/main/java/br/edu/com/tg/manager/core/domain/entities/enums/ProfessorRole.java
@@ -1,4 +1,4 @@
-package br.edu.com.tg.manager.core.domain.entities;
+package br.edu.com.tg.manager.core.domain.entities.enums;
 
 /**
  * Enumeração de domínio.
